+++ test/browser.spec.js
@@ -1,10 +1,11 @@
 // @vitest-environment jsdom
 import { afterEach, beforeEach } from 'vitest'
 import { setup, teardown } from './helpers/setup.js'
+import zlFetch, { createZlFetch } from '../src/index.js'
 
 import integrationTests from './helpers/integration-tests.js'
 
 beforeEach(setup)
 afterEach(teardown)
 
-integrationTests('Browser')
+integrationTests('Browser', { zlFetch, createZlFetch })
