+++ test/node.spec.js
@@ -1,9 +1,10 @@
 import { afterEach, beforeEach } from 'vitest'
 import { setup, teardown } from './helpers/setup.js'
+import zlFetch, { createZlFetch } from '../src/index.js'
 
 import integrationTests from './helpers/integration-tests.js'
 
 beforeEach(setup)
 afterEach(teardown)
 
-integrationTests('Node')
+integrationTests('Node', { zlFetch, createZlFetch })
